fix data lag_matrix data type

+++ pyunicorn/funcnet/coupling_analysis.py
@@ -124,8 +124,7 @@ class CouplingAnalysis(object):
         """
 
         N = self.N
-        return _symmetrize_by_absmax(similarity_matrix,
-                                     lag_matrix.astype(numpy.int), N)
+        return _symmetrize_by_absmax(similarity_matrix, lag_matrix, N)
 
     #
     #  Define methods to estimate similarity measures
